pyproject.toml
[project]
name = "claude-memory"
version = "0.1.0"
description = "Vault-based long-term memory plugin for Claude Code"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.29",
    "python-dotenv>=1.0.0",
    "tzdata>=2024.1",
]

[tool.ruff]
line-length = 100
